// src/app/features/order/models/orderDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface OrderDto { 
    itemsId?: Array<number>;
    restaurantId: number;
    waiterId: number;
    price: number;
    tableId?: number;
    orderType: OrderDto.OrderTypeEnum;
    paymentType: OrderDto.PaymentTypeEnum;
    orderStatus?: OrderDto.OrderStatusEnum;
}
export namespace OrderDto {
    export type OrderTypeEnum = 'DINE_IN' | 'TAKEAWAY' | 'DELIVERY';
    export const OrderTypeEnum = {
        DINEIN: 'DINE_IN' as OrderTypeEnum,
        TAKEAWAY: 'TAKEAWAY' as OrderTypeEnum,
        DELIVERY: 'DELIVERY' as OrderTypeEnum
    };
    export type PaymentTypeEnum = 'CASH' | 'CREDIT_CARD' | 'DEBIT_CARD' | 'DIGITAL_WALLET';
    export const PaymentTypeEnum = {
        CASH: 'CASH' as PaymentTypeEnum,
        CREDITCARD: 'CREDIT_CARD' as PaymentTypeEnum,
        DEBITCARD: 'DEBIT_CARD' as PaymentTypeEnum,
        DIGITALWALLET: 'DIGITAL_WALLET' as PaymentTypeEnum
    };
    export type OrderStatusEnum = 'PENDING' | 'CONFIRMED' | 'READY' | 'DELIVERED' | 'CANCELLED';
    export const OrderStatusEnum = {
        PENDING: 'PENDING' as OrderStatusEnum,
        CONFIRMED: 'CONFIRMED' as OrderStatusEnum,
        READY: 'READY' as OrderStatusEnum,
        DELIVERED: 'DELIVERED' as OrderStatusEnum,
        CANCELLED: 'CANCELLED' as OrderStatusEnum
    };
}